Use YieldFrom in read-csv example
